backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
numpy==2.0.2
pandas==2.2.2
scikit-learn==1.6.1
joblib
mysql-connector-python
python-dotenv
sqlalchemy